Moved Model and Models module outside of geo_near.rb
